Feat : PinService 관련 테스트 코드 작성 및 로직 수정(#24)

+++ project/src/main/java/com/project/domain/pin/api/PinServiceImpl.java
@@ -1,5 +1,6 @@
 package com.project.domain.pin.api;
 
+import com.project.common.exception.BusinessLogicException;
 import com.project.common.exception.EntityNotFoundException;
 import com.project.common.exception.ErrorCode;
 import com.project.common.exception.InvalidValueException;
@@ -17,6 +18,8 @@ import com.project.domain.pin.entity.Pin;
 import com.project.domain.pin.repository.PinRepository;
 import com.project.domain.pinpicture.entity.PinPicture;
 import com.project.domain.pinpicture.repository.PinPictureRepository;
+import com.project.domain.usercircle.entity.UserCircle;
+import com.project.domain.usercircle.repository.UserCircleRepository;
 import com.project.domain.users.entity.Users;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -26,6 +29,9 @@ import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @Service
 @Slf4j
@@ -35,6 +41,7 @@ public class PinServiceImpl implements PinService {
     private final PinRepository pinRepository;
     private final PictureRepository pictureRepository;
     private final CircleRepository circleRepository;
+    private final UserCircleRepository userCircleRepository;
     private final CirclePinRepository circlePinRepository;
     private final PinPictureRepository pinPictureRepository;
     private final S3Uploader s3Uploader;
@@ -47,6 +54,16 @@ public class PinServiceImpl implements PinService {
             Pin pin = request.toEntity();
             Circle circle = circleRepository.findById(circleId).orElseThrow(
                     () -> new EntityNotFoundException("존재하지 않는 써클입니다."));
+
+            // 유저가 가입한 써클에서 핀을 생성할 수 있다.
+            if(userCircleRepository.findByUserIdAndCircleId(user.getId(), circleId).isEmpty()){
+                throw new EntityNotFoundException("가입하지 않은 써클입니다.");
+            }
+
+            if (pictures == null || pictures.isEmpty()) {
+                throw new BusinessLogicException("핀에 저장할 사진이 존재하지 않습니다.",ErrorCode.INVALID_INPUT_VALUE);
+            }
+
 
             locationRepository.save(pin.getLocation());
             pin.setUser(user);
@@ -70,6 +87,11 @@ public class PinServiceImpl implements PinService {
     public PinDTO.PinDetailResponse getPinDetail(Users user, Long pinId) {
         Pin pin = pinRepository.findById(pinId).orElseThrow(
                 () -> new EntityNotFoundException("존재하지 않는 핀입니다."));
+
+        List<Circle> userJoinCircles = userCircleRepository.findAllCircleByUserId(user.getId());
+        if(!isPinCreatedByUser(user, pinId) && !isPinCreatedByCircles(userJoinCircles, pinId)){
+            throw new BusinessLogicException("해당 핀에 대한 접근 권한이 없습니다.", ErrorCode.HANDLE_ACCESS_DENIED);
+        }
 
         List<Picture> pictureList = pinPictureRepository.findAllPicturesByPinId(pinId);
         return new PinDTO.PinDetailResponse(pin, pictureList);
@@ -97,7 +119,6 @@ public class PinServiceImpl implements PinService {
         }).toList();
 
         return new PinDTO.PinDetailListResponse(pinDetailResponseList);
-
     }
 
     @Override
@@ -107,12 +128,18 @@ public class PinServiceImpl implements PinService {
             throw new EntityNotFoundException("존재하지 않는 핀입니다.");
         });
 
-        Location updatedLocation = request.getLocation().toEntity();
-        locationRepository.save(updatedLocation);
-        pin.updateTitle(request.getTitle());
-        pin.updateLocation(updatedLocation);
+        // Title, Location 정보의 변화가 있는가?
+        if (request != null) {
+            Location updatedLocation = request.getLocation().toEntity();
+            locationRepository.save(updatedLocation);
+            pin.updateTitle(request.getTitle());
+            pin.updateLocation(updatedLocation);
+        }
 
-        // 사진 수정
+        // 사진 수정. 새로운 사진 목록에 최소 한 장 이상의 사진이 존재해야 한다.
+        if (pictures.isEmpty()) {
+            throw new BusinessLogicException("핀에 저장할 사진이 존재하지 않습니다.",ErrorCode.INVALID_INPUT_VALUE);
+        }
 
         // PinPicture에서 기존 매핑 삭제하고 새롭게 추가한다.
         pinPictureRepository.deleteAll(pinPictureRepository.findAllByPinId(pinId));
@@ -127,19 +154,23 @@ public class PinServiceImpl implements PinService {
 
     @Override
     public void deletePin(Users user, Long pinId) {
-        // PinPicture, circlePin에서 먼저 삭제한다.
-        // 내가 삭제하면 그룹 내에서도 삭제된다.
-        pinPictureRepository.deleteAll(pinPictureRepository.findAllByPinId(pinId));
-        circlePinRepository.deleteAll(circlePinRepository.findAllByPinId(pinId));
-        pinRepository.delete(
-            pinRepository.findById(pinId).orElseThrow(() -> {
-                log.error("Delete pin failed. pinId = {}", pinId);
-                throw new EntityNotFoundException("존재하지 않는 핀입니다.");
-            })
-        );
+        if (isPinCreatedByUser(user, pinId)) {
+            // PinPicture, circlePin에서 먼저 삭제한다.
+            // 내가 삭제하면 그룹 내에서도 삭제된다.
+            pinPictureRepository.deleteAll(pinPictureRepository.findAllByPinId(pinId));
+            circlePinRepository.deleteAll(circlePinRepository.findAllByPinId(pinId));
+            pinRepository.delete(
+                    pinRepository.findById(pinId).orElseThrow(() -> {
+                        log.error("Delete pin failed. pinId = {}", pinId);
+                        throw new EntityNotFoundException("존재하지 않는 핀입니다.");
+                    })
+            );
+        } else {
+            throw new BusinessLogicException("해당 핀에 대한 접근 권한이 없습니다.", ErrorCode.HANDLE_ACCESS_DENIED);
+        }
     }
 
-    public List<Picture> uploadAndSavePictures(List<MultipartFile> pictureList) {
+    private List<Picture> uploadAndSavePictures(List<MultipartFile> pictureList) {
         return pictureList.stream().map((p) -> {
             // S3에 사진 업로드
             Map<String, String> result = s3Uploader.upload(p, "static");
@@ -149,5 +180,29 @@ public class PinServiceImpl implements PinService {
             // Picture 생성
             return pictureRepository.save(Picture.builder().originalName(pictureName).url(uploadUrl).build());
         }).toList();
+    }
+
+    private boolean isPinCreatedByUser(Users user, Long pinId) {
+        Long userId = user.getId();
+        Pin targetPin = pinRepository.findById(pinId).orElseThrow(() -> {
+            throw new EntityNotFoundException("존재하지 않는 핀입니다.");
+        });
+
+        List<Pin> myPins = pinRepository.findByUserId(userId);
+        return myPins.contains(targetPin);
+    }
+
+    private boolean isPinCreatedByCircles(List<Circle> circles, Long pinId) {
+        Pin targetPin = pinRepository.findById(pinId).orElseThrow(() -> {
+            throw new EntityNotFoundException("존재하지 않는 핀입니다.");
+        });
+
+        for (Circle circle : circles) {
+            List<Pin> pinsByCircle = circlePinRepository.findAllPinsByCircleId(circle.getId());
+            if (pinsByCircle.contains(targetPin)) {
+                return true;
+            }
+        }
+        return false;
     }
 }
